integrate Capybara with Poltergeist and some refactors.
